Modified main.h

+++ main.h
@@ -10,7 +10,8 @@ int _putchar(char c);
 int _printf(const char *format, ...);
 int _printfstr(char *str, int j);
 int _printfcharc(char c, int j);
-int _printfint_double(char c, int i, int j);
-int _printfint(int i, int j);
+int _printfint_double(char c, long int i, int j);
+int _printfint(long int i, int j);
+
 
 #endif
